+++ core/rest-services/src/main/java/com/company/project/webservice/interfaces/base/BaseRestService.java
@@ -24,6 +24,6 @@ public interface BaseRestService<E extends Serializable> extends BaseReadOnlyRes
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	@ResponseStatus(value = HttpStatus.OK)
-	public void delete(Integer id);
+	public void delete(Long id);
 
 }
